Rename all the things! Lower case camel case
