+++ src/components/news/articleBody.tsx
@@ -11,7 +11,7 @@ import { palette } from '@guardian/src-foundations';
 import { from, until } from '@guardian/src-foundations/mq';
 import { render } from "renderBlocks";
 import { BlockElement } from 'types/capi-thrift-models';
-import { PillarStyles } from 'types/Pillar';
+import { PillarStyles } from 'types/pillar';
 
 const richLinkWidth = "13.75rem";
