+++ src/components/news/articleHeadline.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import ArticleRating from './ArticleRating';
+import ArticleRating from './articleRating';
 import { basePx, sidePadding, headlineFont, darkModeCss } from 'styles';
 import { css, SerializedStyles } from '@emotion/core'
 import { palette } from '@guardian/src-foundations'
-import { PillarStyles } from 'types/Pillar';
+import { PillarStyles } from 'types/pillar';
 
 const HeadlineStyles = (feature: boolean, { featureHeadline }: PillarStyles): SerializedStyles => css`
     padding: ${basePx(0, 0, 4, 0)};
