+++ src/components/shared/avatar.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { configure, shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Avatar from 'components/shared/Avatar';
+import Avatar from 'components/shared/avatar';
 
 configure({ adapter: new Adapter() });
