+++ src/renderBlocks.ts
@@ -3,7 +3,7 @@
 import React from 'react';
 import jsdom from 'jsdom';
 
-import { Result, Err, fromUnsafe } from './types/Result';
+import { Result, Err, fromUnsafe } from './types/result';
 import { imageBlock } from './components/blocks/image';
 import { insertAdPlaceholders } from './ads';
 import { transform } from 'utils/contentTransformations';
