+++ src/styles.test.ts
@@ -1,5 +1,5 @@
 import { basePx, baseMultiply, darkModeCss } from './styles';
-import { getPillarStyles, Pillar } from 'types/Pillar';
+import { getPillarStyles, Pillar } from 'types/pillar';
 
 describe('helper functions return correct styles', () => {
     test('Calculates base pixels', () => {
