+++ src/types/functor.ts
